pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "auto-scholar"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.110.0,<1.0.0",
    "uvicorn>=0.27.0,<1.0.0",
    "langgraph>=0.2.0,<1.0.0",
    "langgraph-checkpoint-sqlite>=1.0.0,<2.0.0",
    "langchain-core>=0.3.0,<1.0.0",
    "openai>=1.12.0,<2.0.0",
    "pydantic>=2.5.0,<3.0.0",
    "aiohttp>=3.9.0,<4.0.0",
    "tenacity>=8.2.0,<9.0.0",
    "python-dotenv>=1.0.0,<2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0,<9.0.0",
    "pytest-asyncio>=0.23.0,<1.0.0",
    "pytest-cov>=4.1.0,<5.0.0",
    "httpx>=0.27.0,<1.0.0",
    "ruff>=0.5.0,<1.0.0",
    "mypy>=1.10.0,<2.0.0",
    "pre-commit>=3.7.0,<4.0.0",
]

[tool.setuptools.packages.find]
include = ["backend*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests requiring external APIs",
]

[tool.coverage.run]
branch = true
source = ["backend"]
omit = ["backend/__init__.py"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "raise NotImplementedError",
]
